+++ gnat/src/pipeline/sample.rs
@@ -19,9 +19,9 @@ use crate::pipeline::Interval;
 use crate::pipeline::StreamType;
 use crate::utils::duckdb::duckdb_open_memory;
 use chrono::{DateTime, Utc};
-
 use std::fs;
 use std::io::Error;
+use std::path::Path;
 
 pub struct SampleProcessor {
     pub command: String,
@@ -66,6 +66,13 @@ impl SampleProcessor {
             .parse::<u8>()
             .unwrap();
 
+        if interval == Interval::MINUTE || interval == Interval::SECOND {          
+            return Err(Error::new(
+                std::io::ErrorKind::InvalidInput,
+                "sampling not supported for minute or second intervals",
+            ));
+        }
+
         let mut input_list = Vec::<String>::new();
         input_list.push(input.to_string());
         let mut output_list = Vec::<String>::new();
@@ -83,7 +90,7 @@ impl SampleProcessor {
     }
 
     fn purge_old(&self) -> Result<(), Error> {
-        let conn = duckdb_open_memory(2);
+
         let current_utc: DateTime<Utc> = Utc::now();
         let rfc3339_name: String = current_utc.to_rfc3339();
         let new_filename = format!(
@@ -93,7 +100,7 @@ impl SampleProcessor {
         );
         let tmp_filename = format!("{}/.{}", self.output_list[0], new_filename);
         let final_filename = format!("{}/{}.parquet", self.output_list[0], new_filename);
-
+            
         let mut file_list: Vec<String> = Vec::new();
         for entry in fs::read_dir(&self.output_list[0])
             .map_err(|e| Error::new(std::io::ErrorKind::Other, format!("read_dir error: {}", e)))?
@@ -111,10 +118,10 @@ impl SampleProcessor {
         }
 
         if file_list.is_empty() {
-            println!("{}: no files to purge", self.command);
+            println!("{}: no files to filter", self.command);
             return Ok(());
         }
-        
+
         let parquet_list = file_list
             .iter()
             .map(|file| format!("'{}'", file))
@@ -122,32 +129,40 @@ impl SampleProcessor {
             .join(",");
         let parquet_list = format!("[{}]", parquet_list);
 
+        let conn = duckdb_open_memory(4);           
         let sql_command = format!(
             "COPY (SELECT * FROM read_parquet({})
                    WHERE date_trunc('day',stime) > date_add(date_trunc('day',stime), - INTERVAL {} DAY))
                 TO '{}' (FORMAT 'parquet', CODEC 'snappy', ROW_GROUP_SIZE 100_000);",
             parquet_list, self.retention, tmp_filename
         );
-        //println!("{}: {}", self.command,sql_command);
+        
+        println!("{}: filtering...", self.command);
         conn.execute_batch(&sql_command)
             .map_err(|e| Error::new(std::io::ErrorKind::Other, format!("DuckDB error: {}", e)))?;
+        println!("{}: done.", self.command);
 
         for file in &file_list {
-            fs::remove_file(file).map_err(|e| {
-                Error::new(
-                    std::io::ErrorKind::Other,
-                    format!("failed to remove file {}: {}", file, e),
-                )
+            if Path::new(&tmp_filename).exists() {
+                fs::remove_file(file).map_err(|e| {
+                    Error::new(
+                        std::io::ErrorKind::Other,
+                        format!("failed to remove file {}: {}", file, e),
+                    )
+                })?;
+            }
+        }
+
+        if Path::new(&tmp_filename).exists() {
+            fs::rename(&tmp_filename, &final_filename).map_err(|e| {
+                Error::new(std::io::ErrorKind::Other, format!("renaming error: {}", e))
             })?;
         }
 
-        fs::rename(&tmp_filename, &final_filename)
-            .map_err(|e| Error::new(std::io::ErrorKind::Other, format!("renaming error: {}", e)))?;
-        println!("{}: purged old records.", self.command);
         Ok(())
     }
 
-    fn generate_samples(&mut self, file_list: &Vec<String>) -> Result<(), Error> {
+    fn process_samples(&mut self, file_list: &Vec<String>) -> Result<i64, Error> {
         // Use iterator and join for file list formatting
         let parquet_list = file_list
             .iter()
@@ -167,37 +182,10 @@ impl SampleProcessor {
                     "{}: invalid stream of parquet files, skipping",
                     self.command
                 );
-                return Ok(());
+                return Ok(0);
             }
         }
         let conn = duckdb_open_memory(2);
-        // check the number of days in the dataset
-        /*
-        {
-            println!("{}: checking dataset duration...", self.command);
-            // Use date_diff to calculate the number of days between the first and last timestamps
-            let sql_days_command = format!(
-                "SELECT date_diff('day',first,last)
-                 FROM (SELECT MIN(stime) AS first, MAX(stime) AS last
-                 FROM read_parquet({}));",
-                parquet_list
-            );
-            let mut stmt = conn.prepare(&sql_days_command).map_err(|e| {
-                Error::new(
-                    std::io::ErrorKind::Other,
-                    format!("DuckDB prepare error: {}", e),
-                )
-            })?;
-            let days = stmt
-                .query_row([], |row| Ok(row.get::<_, u32>(0).expect("missing version")))
-                .expect("missing days");
-            println!("{}: {} days of data", self.command, days);
-            if days < MINIMUM_DAYS {
-                println!("{}: not enough data to sample, skipping", self.command);
-                return Ok(());
-            }
-        }
-        */
 
         let sql_command = format!(
             "CREATE TABLE flow AS SELECT * FROM read_parquet({});",
@@ -226,6 +214,7 @@ impl SampleProcessor {
                 Error::new(std::io::ErrorKind::Other, format!("query map error: {}", e))
             })?;
 
+        let mut records_processed = 0;
         for record in record_iter {
             let record = record.map_err(|e| {
                 Error::new(std::io::ErrorKind::Other, format!("record error: {}", e))
@@ -234,6 +223,30 @@ impl SampleProcessor {
                 "{}: sampling [{}/{}/{}] {}%",
                 self.command, record.observe, record.vlan, record.proto, self.percent
             );
+
+            // Check if there are any records for this observation
+            // If there are no records, skip to the next observation
+            // This is a performance optimization to avoid processing empty observations
+            let sql_command = format!(
+                "SELECT count(*) FROM flow
+                 WHERE observe='{}' 
+                   AND dvlan = {} AND proto='{}'                   
+                   AND TRIGGER = 0;",
+                record.observe, record.vlan, record.proto
+            );
+
+            let mut stmt = conn.prepare(&sql_command).map_err(|e| {
+                Error::new(std::io::ErrorKind::Other, format!("DuckDB error: {}", e))
+            })?;
+            let record_count: i64 = stmt.query_row([], |row| row.get(0)).map_err(|e| {
+                Error::new(std::io::ErrorKind::Other, format!("DuckDB error: {}", e))
+            })?;
+
+            if record_count == 0 {
+                continue; // Skip if no records found for this observation
+            }
+
+            // Proceed with sampling if records are found
             let current_utc: DateTime<Utc> = Utc::now();
             let rfc3339_name: String = current_utc.to_rfc3339();
             let new_filename = format!(
@@ -258,12 +271,17 @@ impl SampleProcessor {
             conn.execute_batch(&sql_command).map_err(|e| {
                 Error::new(std::io::ErrorKind::Other, format!("DuckDB error: {}", e))
             })?;
-            fs::rename(&tmp_filename, &final_filename).map_err(|e| {
-                Error::new(std::io::ErrorKind::Other, format!("renaming error: {}", e))
-            })?;
+
+            if Path::new(&tmp_filename).exists() {
+                fs::rename(&tmp_filename, &final_filename).map_err(|e| {
+                    Error::new(std::io::ErrorKind::Other, format!("renaming error: {}", e))
+                })?;
+            }
+
+            records_processed += record_count;
         }
         println!("{}: sampled records.", self.command);
-        Ok(())
+        Ok(records_processed)
     }
 }
 
@@ -298,8 +316,14 @@ impl FileProcessor for SampleProcessor {
         true
     }
     fn process(&mut self, file_list: &Vec<String>) -> Result<(), Error> {
-        self.generate_samples(file_list)?;
-        self.purge_old()?;
+        if self
+            .process_samples(file_list)
+            .expect("process_samples failed")
+            > 0
+        {
+            self.purge_old()?;
+        }
+
         Ok(())
     }
 }
